Move githhubEventHelpers to api/github/helpers
