+++ src/api/github/helpers.ts
@@ -3,22 +3,6 @@ import * as Sentry from '@sentry/node';
 
 import { GitHubOrg } from '@api/github/org';
 
-// Validation Helpers
-
-export async function shouldSkip(payload, org: GitHubOrg, reasonsToSkip) {
-  // Could do Promise-based async here, but that was getting complicated[1] and
-  // there's not really a performance concern (famous last words).
-  //
-  // [1] https://github.com/getsentry/eng-pipes/pull/212#discussion_r657365585
-
-  for (const skipIf of reasonsToSkip) {
-    if (await skipIf(payload, org)) {
-      return true;
-    }
-  }
-  return false;
-}
-
 async function sendQuery(query: string, data: object, octokit: Octokit) {
   let response: any;
   try {
