+++ src/utils/shouldSkip.ts
@@ -1,0 +1,10 @@
+import { GitHubOrg } from '@api/github/org';
+
+export async function shouldSkip(payload, org: GitHubOrg, reasonsToSkip) {
+  for (const skipIf of reasonsToSkip) {
+    if (await skipIf(payload, org)) {
+      return true;
+    }
+  }
+  return false;
+}
